models/StyleGAN2/requirements.txt
jittor==1.3.2.5
pillow
pyyaml==6.0
tqdm==4.46.0
wandb==0.12.19
munch==2.5.0
numpy==1.18.2
